refactor: split reporter_memory into store and search packages
Break up the large context_store and search modules so persistence and
retrieval can evolve independently without changing the public API.
